Advance TS checker + harness
- Fix determinism and query semantics in typecheck-ts\n- Improve HIR lowering for TS parameter properties\n- Fix BigInt literal emission in emit-js\n- Make difftsc baselines act as assertions for type facts\n- Vendor/pin TypeScript for the harness (npm ci) and add robust loader/probe

+++ typecheck-ts-cli/src/main.rs
@@ -142,7 +142,7 @@ struct TypeAtResult {
 struct SymbolAtResult {
   file: String,
   offset: u32,
-  symbol: u32,
+  symbol: u64,
   #[serde(skip_serializing_if = "Option::is_none")]
   name: Option<String>,
   #[serde(skip_serializing_if = "Option::is_none")]
@@ -156,7 +156,7 @@ struct SymbolAtResult {
 
 #[derive(Serialize)]
 struct ExportEntryJson {
-  symbol: u32,
+  symbol: u64,
   #[serde(skip_serializing_if = "Option::is_none")]
   def: Option<u32>,
   #[serde(skip_serializing_if = "Option::is_none")]
